requirements.txt
tensorflow
keras
gradio
opencv-python
numpy
pillow
matplotlib
scikit-learn
reportlab